+++ test/bootstrap/node.js
@@ -1,3 +1,0 @@
-import {buildGetJSON} from '../../src/index-node.mjs';
-
-global.buildGetJSON = buildGetJSON;